pyproject.toml
[tool.poetry]
name = "drfproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "^3.8.1"
django = "^5.1.3"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
sqlparse = "^0.5.1"
tzdata = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
